Time Series enhancements: Fixed ETK2 build error (py2->py3); refactored TimeSeriesRegion into new file; added new measurement representation and UID generation to extraction; renamed all files to be consistent with the naming conventions
