+++ etk/timeseries_processor.py
@@ -1,4 +1,4 @@
-import etk.timeseries.extractSpreadsheet as extractSpreadsheet
+import etk.timeseries.extract_spreadsheet as extractSpreadsheet
 from typing import List
 from etk.document import Document
 from etk.etk_exceptions import InvalidArgumentsError
@@ -32,7 +32,7 @@ class TimeseriesProcessor(object):
                          data_set: str = None,) -> List[Document]:
         documents = list()
 
-        for json_obj in json_objs[0]:
+        for json_obj in json_objs:
             cdr_doc = json_obj
 
             if file_name is not None:
