+++ examples/timeseries_processor/resources/DIESEL_june_annotation.json
@@ -8,7 +8,7 @@
 			{
 				"orientation": "row",
 				"rows": "[5:41]",
-				"locs": "[J:S]",
+				"locs": "[B:S]",
 				"metadata": [
 					{
 						"source": "col",
